Refactored LayerListItem and Layer's default color
