+++ web-ng/src/app/components/feature-panel/feature-panel.module.ts
@@ -24,6 +24,7 @@ import { MatListModule } from '@angular/material/list';
 import { MatMenuModule } from '@angular/material/menu';
 
 import { LayerListItemModule } from '../layer-list-item/layer-list-item.module';
+import { FeatureDetailsModule } from '../feature-details/feature-details.module';
 
 @NgModule({
   imports: [
@@ -34,6 +35,7 @@ import { LayerListItemModule } from '../layer-list-item/layer-list-item.module';
     MatIconModule,
     MatButtonModule,
     MatMenuModule,
+    FeatureDetailsModule,
   ],
   exports: [FeaturePanelComponent],
   declarations: [FeaturePanelComponent],
