+++ web-ng/src/app/components/layer-dialog/layer-dialog.component.html
@@ -66,17 +66,7 @@
       <span class="link-add-question">Add question</span>
     </button>
   </mat-dialog-content>
-  <mat-dialog-actions fxLayout fxLayoutAlign="space-between">
-    <div>
-      <button
-        fxFlex="none"
-        mat-icon-button
-        type="button"
-        (click)="onDeleteLayer()"
-      >
-        <mat-icon>delete_forever</mat-icon>
-      </button>
-    </div>
+  <mat-dialog-actions fxLayout fxLayoutAlign="end">
     <div fxLayoutAlign="end">
       <button
         type="button"
