--- web-ng/src/app/components/layer-list-item/layer-list-item.component.css
@@ -1,28 +1,0 @@
-/**
- * Copyright 2019 Google LLC
- *
- * Licensed under the Apache License, Version 2.0 (the 'License');
- * you may not use this file except in compliance with the License.
- * You may obtain a copy of the License at
- *
- *     https://www.apache.org/licenses/LICENSE-2.0
- *
- * Unless required by applicable law or agreed to in writing, software
- * distributed under the License is distributed on an 'AS IS' BASIS,
- * WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
- * See the License for the specific language governing permissions and
- * limitations under the License.
- */
-
-/*
-  Workaround to eliminate double-padding between right-hand side overflow
-  menu control and list item.
-*/
-:host ::ng-deep .mat-list-item-content {
-  padding-right: 8px !important;
-}
-
-.layersIcon {
-  vertical-align: middle;
-  padding-right: 5px;
-}
